Reformat Comment (by way of rubywrap.rb) no longer adds extra whitespace
Previously, using Reformat Comment on a block that looked like

// lorem ...
//
// lorem ...

would add an extra space at the end of the second line every time you ran the command. This was due to the @@re_comment regular expression matching the newline character at the end of the second line and thereby setting the post_indent variable incorrectly. The regexp now only matches trailing spaces after the comment prefix. (Matching tabs is unnecessary because the text is detabified.)