Don't depend on RDoc::Usage any more
This was removed from Ruby 1.9 with no replacement. Since all it
does is print out a nice usage message, and since this is being
used internally by Textmate, AND since this code has not been touched
in 5 years, I think it's perfectly valid to simple hard-wire the
"usage" output and remove the dependency, so that you can reflow
comments when using Ruby 1.9.