New “Strip Wrap Characters” action bound to ‘]’
This will strip a single pair of leading/trailing braces, quotes, and similar. It also works to strip an Objective-C string: @"foo" → foo.

The command is only active when there is a selection.

Presently no effort is done to actually ensure the stripped characters are actually balanced. It doesn’t seem like any real benefit would be derived for adding this complexity.