Don't enable spell checking for strings in embedded code.
We enable spell checking for strings in the source scope which allows sensible spell checking. However inside of the text scope we want to not enable these when they appear in embedded source code (eg: PHP) as it is more common to want to have spell checking enabled in text files without getting it in the source portion.