Support underscore prefix when toggling _camelCase to _snake_case
The support is for arbitrary prefixes but since cycling from snake_case would eat the underscores, it meant that previously the transformation could not be undone by continuing the cycling.

This commit changes the ruby requirement to 2.0 but almost 99% of TextMate users are now on OS X 10.9 or higher (according to software update checks).